added popup component
